requirements.txt
pymupdf
sentence-transformers
faiss-cpu
chromadb
python-dotenv
langchain
langchain-community
langchain-google-genai
chromadb
pypdf
bs4
python-dotenv
tiktoken
unstructured
networkx
langchain-chroma>=0.1.2